requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
deprecation==2.1.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
dotenv==0.9.9
gotrue==2.12.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.2
hyperframe==6.1.0
idna==3.11
packaging==25.0
pillow==12.0.0
postgrest==0.16.11
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
realtime==1.0.6
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.7.7
StrEnum==0.4.15
supabase==2.5.0
supafunc==0.4.7
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==12.0
django_extensions==3.2.3
